fix: share live desktop membership with workspace open groups
