+++ Sources/Cloud/CloudTreeRemoteWorkspaces.swift
@@ -74,28 +74,12 @@ extension CloudTreeNodeBuilder {
                 byWorkspace[workspace.id] = members
             }
         }
-        for member in localDisplayMembers(resources: resources, projections: projections) {
+        for member in SurfaceProjection.localDisplayMembers(resources: resources, projections: projections) {
             var members = byWorkspace[member.workspaceID] ?? .none
             members.displays.append(member.resource)
             byWorkspace[member.workspaceID] = members
         }
         return byWorkspace
-    }
-
-    /// Explicit VNC panes have no daemon tab. Their live catalog projections name
-    /// their bound workspace; availability in the machine display pool never does.
-    /// A daemon placement of the same display already supplies that workspace row.
-    static func localDisplayMembers(resources: [SurfaceResource], projections: [SurfaceProjection]) -> [(resource: SurfaceResource, workspaceID: String)] {
-        let displays = Dictionary(uniqueKeysWithValues: resources.filter { $0.kind == .display }.map { ($0.id, $0) })
-        var seen: [SurfaceResourceID: Set<String>] = [:]
-        return projections.compactMap { projection in
-            guard projection.remoteTabID == nil,
-                  let workspaceID = projection.remoteWorkspaceID,
-                  let resource = displays[projection.resource],
-                  !resource.remoteWorkspaces.contains(where: { $0.id == workspaceID }),
-                  seen[resource.id, default: []].insert(workspaceID).inserted else { return nil }
-            return (resource, workspaceID)
-        }
     }
 
     /// The members of one workspace (an existing workspace nothing views has none).
